RUMM-3014: Move OkHttp instrumentation code into the dedicated module (cannot be compiled yet)
