+++ dd-sdk-android-glide/src/test/kotlin/com/datadog/android/glide/DatadogGlideModuleTest.kt
@@ -15,7 +15,7 @@ import com.bumptech.glide.load.engine.executor.GlideExecutor
 import com.bumptech.glide.load.model.GlideUrl
 import com.bumptech.glide.load.model.ModelLoaderFactory
 import com.bumptech.glide.module.AppGlideModule
-import com.datadog.android.DatadogInterceptor
+import com.datadog.android.okhttp.DatadogInterceptor
 import com.datadog.tools.unit.assertj.containsInstanceOf
 import com.datadog.tools.unit.getFieldValue
 import com.nhaarman.mockitokotlin2.argumentCaptor
