+++ dd-sdk-android-okhttp/src/test/kotlin/com/datadog/android/okhttp/rum/RumInterceptorTest.kt
@@ -4,12 +4,13 @@
  * Copyright 2016-Present Datadog, Inc.
  */
 
-package com.datadog.android.rum
+package com.datadog.android.okhttp.rum
 
 import com.datadog.android.core.sampling.RateBasedSampler
+import com.datadog.android.okhttp.tracing.TracingInterceptor
+import com.datadog.android.okhttp.tracing.TracingInterceptorNotSendingSpanTest
+import com.datadog.android.rum.NoOpRumResourceAttributesProvider
 import com.datadog.android.tracing.NoOpTracedRequestListener
-import com.datadog.android.tracing.TracingInterceptor
-import com.datadog.android.tracing.TracingInterceptorNotSendingSpanTest
 import com.datadog.android.utils.config.GlobalRumMonitorTestConfiguration
 import com.datadog.android.utils.forge.Configurator
 import com.datadog.tools.unit.annotations.TestConfigurationsProvider
