+++ dd-sdk-android-okhttp/src/test/kotlin/com/datadog/android/okhttp/tracing/TracingInterceptorNonDdTracerNotSendingSpanTest.kt
@@ -4,21 +4,23 @@
  * Copyright 2016-Present Datadog, Inc.
  */
 
-package com.datadog.android.tracing
+package com.datadog.android.okhttp.tracing
 
+import android.content.Context
 import android.util.Log
 import com.datadog.android.Datadog
 import com.datadog.android.core.configuration.Configuration
 import com.datadog.android.core.internal.net.FirstPartyHostHeaderTypeResolver
 import com.datadog.android.core.internal.utils.loggableStackTrace
-import com.datadog.android.core.sampling.RateBasedSampler
 import com.datadog.android.core.sampling.Sampler
+import com.datadog.android.tracing.TracingHeaderType
+import com.datadog.android.utils.config.ApplicationContextTestConfiguration
+import com.datadog.android.utils.config.CoreFeatureTestConfiguration
 import com.datadog.android.utils.config.InternalLoggerTestConfiguration
 import com.datadog.android.utils.forge.Configurator
 import com.datadog.android.v2.api.InternalLogger
 import com.datadog.android.v2.core.DatadogCore
 import com.datadog.android.v2.core.NoOpSdkCore
-import com.datadog.opentracing.DDSpanContext
 import com.datadog.opentracing.DDTracer
 import com.datadog.tools.unit.annotations.TestConfigurationsProvider
 import com.datadog.tools.unit.extensions.TestConfigurationExtension
@@ -74,6 +76,10 @@ import java.util.Locale
 import java.util.concurrent.CountDownLatch
 import java.util.concurrent.TimeUnit
 
+/**
+ * This test is a duplicate of [TracingInterceptorNotSendingSpanTest] but assuming the Tracer is
+ * not our own implementation and therefore doesn't implement DD specific methods.
+ */
 @Extensions(
     ExtendWith(MockitoExtension::class),
     ExtendWith(ForgeExtension::class),
@@ -81,7 +87,7 @@ import java.util.concurrent.TimeUnit
 )
 @MockitoSettings(strictness = Strictness.LENIENT)
 @ForgeConfiguration(Configurator::class)
-internal open class TracingInterceptorTest {
+internal open class TracingInterceptorNonDdTracerNotSendingSpanTest {
 
     lateinit var testedInterceptor: TracingInterceptor
 
@@ -94,12 +100,12 @@ internal open class TracingInterceptorTest {
     lateinit var mockLocalTracer: Tracer
 
     @Mock
-    lateinit var mockSpanBuilder: DDTracer.DDSpanBuilder
+    lateinit var mockSpanBuilder: Tracer.SpanBuilder
 
     @Mock
-    lateinit var mockSpanContext: DDSpanContext
+    lateinit var mockSpanContext: SpanContext
 
-    @Mock(extraInterfaces = [MutableSpan::class])
+    @Mock
     lateinit var mockSpan: Span
 
     @Mock
@@ -118,6 +124,12 @@ internal open class TracingInterceptorTest {
 
     // region Fakes
 
+    @StringForgery(regex = HOSTNAME_PATTERN)
+    lateinit var fakeHostName: String
+
+    @StringForgery(regex = IPV4_PATTERN)
+    lateinit var fakeHostIp: String
+
     lateinit var fakeMethod: String
     var fakeBody: String? = null
     var fakeMediaType: MediaType? = null
@@ -126,8 +138,6 @@ internal open class TracingInterceptorTest {
     lateinit var fakeResponseBody: String
 
     lateinit var fakeUrl: String
-
-    lateinit var fakeBaseUrl: String
 
     @Forgery
     lateinit var fakeConfig: Configuration.Feature.Tracing
@@ -149,12 +159,10 @@ internal open class TracingInterceptorTest {
     // endregion
 
     @BeforeEach
-    fun `set up`(forge: Forge) {
+    open fun `set up`(forge: Forge) {
         Datadog.setVerbosity(Log.VERBOSE)
 
         whenever(mockTracer.buildSpan(TracingInterceptor.SPAN_NAME)) doReturn mockSpanBuilder
-        whenever(mockLocalTracer.buildSpan(TracingInterceptor.SPAN_NAME)) doReturn mockSpanBuilder
-        whenever(mockSpanBuilder.withOrigin(fakeOrigin)) doReturn mockSpanBuilder
         whenever(mockSpanBuilder.asChildOf(null as SpanContext?)) doReturn mockSpanBuilder
         whenever(mockSpanBuilder.start()) doReturn mockSpan
         whenever(mockSpan.context()) doReturn mockSpanContext
@@ -162,105 +170,63 @@ internal open class TracingInterceptorTest {
         whenever(mockSpanContext.toTraceId()) doReturn fakeTraceId
         whenever(mockTraceSampler.sample()) doReturn true
 
-        val mediaType = forge.anElementFrom("application", "image", "text", "model") +
-            "/" + forge.anAlphabeticalString()
-        fakeLocalHosts =
-            forge.aMap { forge.aStringMatching(HOSTNAME_PATTERN) to setOf(TracingHeaderType.DATADOG) }
-        fakeMediaType = MediaType.parse(mediaType)
-        fakeUrl = forgeUrlWithQueryParams(forge)
+        fakeMediaType = if (forge.aBool()) {
+            val mediaType = forge.anElementFrom("application", "image", "text", "model") +
+                "/" + forge.anAlphabeticalString()
+            MediaType.parse(mediaType)
+        } else {
+            null
+        }
+        fakeLocalHosts = forge.aMap {
+            forge.aStringMatching(TracingInterceptorTest.HOSTNAME_PATTERN) to setOf(
+                TracingHeaderType.DATADOG
+            )
+        }
+        fakeUrl = forgeUrl(forge)
         fakeRequest = forgeRequest(forge)
         val mockCore = mock<DatadogCore>()
         whenever(mockCore.tracingFeature) doReturn mock()
         Datadog.globalSdkCore = mockCore
-        testedInterceptor = instantiateTestedInterceptor(fakeLocalHosts) {
-            mockLocalTracer
-        }
+        doAnswer { false }.whenever(mockResolver).isFirstPartyUrl(any<HttpUrl>())
+        doAnswer { true }.whenever(mockResolver).isFirstPartyUrl(HttpUrl.get(fakeUrl))
 
         GlobalTracer.registerIfAbsent(mockTracer)
-    }
-
-    open fun instantiateTestedInterceptor(
-        tracedHosts: Map<String, Set<TracingHeaderType>> = emptyMap(),
-        factory: (Set<TracingHeaderType>) -> Tracer
-    ): TracingInterceptor {
-        return TracingInterceptor(
-            tracedHosts,
-            mockRequestListener,
-            mockResolver,
-            fakeOrigin,
-            mockTraceSampler,
-            factory
-        )
-    }
-
-    open fun getExpectedOrigin(): String {
-        return fakeOrigin
+        testedInterceptor = instantiateTestedInterceptor(fakeLocalHosts) { mockLocalTracer }
     }
 
     @AfterEach
-    fun `tear down`() {
-        Datadog.globalSdkCore = NoOpSdkCore()
+    open fun `tear down`() {
         GlobalTracer::class.java.setStaticValue("isRegistered", false)
+        Datadog.globalSdkCore = NoOpSdkCore()
     }
 
-    @Test
-    fun `M instantiate with default values W init()`() {
-        // When
-        val interceptor = TracingInterceptor()
-
-        // Then
-        assertThat(interceptor.tracedHosts).isEmpty()
-        assertThat(interceptor.tracedRequestListener)
-            .isInstanceOf(NoOpTracedRequestListener::class.java)
-        assertThat(interceptor.traceSampler)
-            .isInstanceOf(RateBasedSampler::class.java)
-        val traceSampler = interceptor.traceSampler as RateBasedSampler
-        assertThat(traceSampler.sampleRate).isEqualTo(
-            TracingInterceptor.DEFAULT_TRACE_SAMPLING_RATE / 100
-        )
-    }
-
-    @Test
-    fun `M instantiate with default values W init()`(
-        @StringForgery(regex = "[a-z]+\\.[a-z]{3}") hosts: List<String>
-    ) {
-        // When
-        val interceptor = TracingInterceptor(hosts)
-
-        // Then
-        assertThat(interceptor.tracedHosts.keys).containsAll(hosts)
-        assertThat(interceptor.tracedRequestListener)
-            .isInstanceOf(NoOpTracedRequestListener::class.java)
-        assertThat(interceptor.traceSampler)
-            .isInstanceOf(RateBasedSampler::class.java)
-        val traceSampler = interceptor.traceSampler as RateBasedSampler
-        assertThat(traceSampler.sampleRate).isEqualTo(
-            TracingInterceptor.DEFAULT_TRACE_SAMPLING_RATE / 100
-        )
+    open fun instantiateTestedInterceptor(
+        tracedHosts: Map<String, Set<TracingHeaderType>>,
+        factory: (Set<TracingHeaderType>) -> Tracer
+    ): TracingInterceptor {
+        return object :
+            TracingInterceptor(
+                tracedHosts,
+                mockRequestListener,
+                mockResolver,
+                fakeOrigin,
+                mockTraceSampler,
+                factory
+            ) {
+            override fun canSendSpan(): Boolean {
+                return false
+            }
+        }
     }
 
     @Test
     fun `𝕄 inject tracing header 𝕎 intercept() {global known host}`(
         @StringForgery key: String,
         @StringForgery(type = StringForgeryType.ALPHA_NUMERICAL) value: String,
-        @IntForgery(min = 200, max = 600) statusCode: Int,
-        forge: Forge
+        @IntForgery(min = 200, max = 600) statusCode: Int
     ) {
-        stubChain(mockChain, statusCode)
         whenever(mockResolver.isFirstPartyUrl(HttpUrl.get(fakeUrl))).thenReturn(true)
-        whenever(mockResolver.headerTypesForUrl(HttpUrl.get(fakeUrl))).thenReturn(
-            setOf(
-                forge.anElementFrom(
-                    setOf(
-                        TracingHeaderType.DATADOG,
-                        TracingHeaderType.B3,
-                        TracingHeaderType.B3MULTI,
-                        TracingHeaderType.TRACECONTEXT
-                    )
-                )
-            )
-        )
-
+        stubChain(mockChain, statusCode)
         doAnswer { invocation ->
             val carrier = invocation.arguments[2] as TextMapInject
             carrier.put(key, value)
@@ -271,12 +237,12 @@ internal open class TracingInterceptorTest {
         assertThat(response).isSameAs(fakeResponse)
         argumentCaptor<Request> {
             verify(mockChain).proceed(capture())
-            assertThat(firstValue.headers(key)).containsOnly(value)
+            assertThat(lastValue.header(key)).isEqualTo(value)
         }
     }
 
     @Test
-    fun `𝕄 inject non-tracing datadog header 𝕎 intercept() {global known host + not sampled}`(
+    fun `𝕄 inject non-tracing header 𝕎 intercept() {global known host + not sampled}`(
         @IntForgery(min = 200, max = 600) statusCode: Int
     ) {
         // Given
@@ -300,7 +266,6 @@ internal open class TracingInterceptorTest {
                 .isEqualTo("0")
             assertThat(lastValue.header(TracingInterceptor.DATADOG_TRACE_ID_HEADER)).isNull()
             assertThat(lastValue.header(TracingInterceptor.DATADOG_SPAN_ID_HEADER)).isNull()
-            assertThat(lastValue.header(TracingInterceptor.DATADOG_ORIGIN_HEADER)).isNull()
         }
     }
 
@@ -390,58 +355,13 @@ internal open class TracingInterceptorTest {
     }
 
     @Test
-    fun `𝕄 inject all non-tracing header 𝕎 intercept() {global known host + not sampled}`(
-        @IntForgery(min = 200, max = 600) statusCode: Int
-    ) {
-        // Given
-        whenever(mockTraceSampler.sample()).thenReturn(false)
-        whenever(mockResolver.isFirstPartyUrl(HttpUrl.get(fakeUrl))).thenReturn(true)
-        whenever(mockResolver.headerTypesForUrl(HttpUrl.get(fakeUrl))).thenReturn(
-            setOf(
-                TracingHeaderType.DATADOG,
-                TracingHeaderType.B3,
-                TracingHeaderType.B3MULTI,
-                TracingHeaderType.TRACECONTEXT
-            )
-        )
-        stubChain(mockChain, statusCode)
-
-        // When
-        val response = testedInterceptor.intercept(mockChain)
-
-        // Then
-        assertThat(response).isSameAs(fakeResponse)
-        argumentCaptor<Request> {
-            verify(mockChain).proceed(capture())
-            assertThat(lastValue.header(TracingInterceptor.DATADOG_SAMPLING_PRIORITY_HEADER))
-                .isEqualTo("0")
-            assertThat(lastValue.header(TracingInterceptor.DATADOG_TRACE_ID_HEADER)).isNull()
-            assertThat(lastValue.header(TracingInterceptor.DATADOG_SPAN_ID_HEADER)).isNull()
-            assertThat(lastValue.header(TracingInterceptor.DATADOG_ORIGIN_HEADER)).isNull()
-            assertThat(lastValue.header(TracingInterceptor.B3M_SAMPLING_PRIORITY_KEY))
-                .isEqualTo("0")
-            assertThat(lastValue.header(TracingInterceptor.B3M_SPAN_ID_KEY)).isNull()
-            assertThat(lastValue.header(TracingInterceptor.B3M_TRACE_ID_KEY)).isNull()
-            assertThat(lastValue.header(TracingInterceptor.B3_HEADER_KEY))
-                .isEqualTo("0")
-            assertThat(lastValue.header(TracingInterceptor.W3C_TRACEPARENT_KEY))
-                .isEqualTo(
-                    "00-%s-%s-00".format(
-                        mockSpan.context().toTraceId(),
-                        mockSpan.context().toSpanId()
-                    )
-                )
-        }
-    }
-
-    @Test
     fun `𝕄 inject tracing header 𝕎 intercept() {local known host}`(
         @StringForgery key: String,
         @StringForgery(type = StringForgeryType.ALPHA_NUMERICAL) value: String,
         @IntForgery(min = 200, max = 600) statusCode: Int,
         forge: Forge
     ) {
-        fakeUrl = forgeUrlWithQueryParams(forge, forge.anElementFrom(fakeLocalHosts.keys))
+        fakeUrl = forgeUrl(forge, forge.anElementFrom(fakeLocalHosts.keys))
         fakeRequest = forgeRequest(forge)
         whenever(mockResolver.isFirstPartyUrl(HttpUrl.get(fakeUrl))).thenReturn(false)
         stubChain(mockChain, statusCode)
@@ -455,7 +375,7 @@ internal open class TracingInterceptorTest {
         assertThat(response).isSameAs(fakeResponse)
         argumentCaptor<Request> {
             verify(mockChain).proceed(capture())
-            assertThat(firstValue.headers(key)).containsOnly(value)
+            assertThat(lastValue.header(key)).isEqualTo(value)
         }
     }
 
@@ -466,7 +386,7 @@ internal open class TracingInterceptorTest {
     ) {
         // Given
         whenever(mockTraceSampler.sample()).thenReturn(false)
-        fakeUrl = forgeUrlWithQueryParams(forge, forge.anElementFrom(fakeLocalHosts.keys))
+        fakeUrl = forgeUrl(forge, forge.anElementFrom(fakeLocalHosts.keys))
         fakeRequest = forgeRequest(forge)
         whenever(mockResolver.isFirstPartyUrl(HttpUrl.get(fakeUrl))).thenReturn(false)
         stubChain(mockChain, statusCode)
@@ -482,7 +402,6 @@ internal open class TracingInterceptorTest {
                 .isEqualTo("0")
             assertThat(lastValue.header(TracingInterceptor.DATADOG_TRACE_ID_HEADER)).isNull()
             assertThat(lastValue.header(TracingInterceptor.DATADOG_SPAN_ID_HEADER)).isNull()
-            assertThat(lastValue.header(TracingInterceptor.DATADOG_ORIGIN_HEADER)).isNull()
         }
     }
 
@@ -493,10 +412,13 @@ internal open class TracingInterceptorTest {
     ) {
         // Given
         whenever(mockTraceSampler.sample()).thenReturn(false)
-        fakeLocalHosts =
-            forge.aMap { forge.aStringMatching(HOSTNAME_PATTERN) to setOf(TracingHeaderType.B3MULTI) }
+        fakeLocalHosts = forge.aMap {
+            forge.aStringMatching(TracingInterceptorTest.HOSTNAME_PATTERN) to setOf(
+                TracingHeaderType.B3MULTI
+            )
+        }
         testedInterceptor = instantiateTestedInterceptor(fakeLocalHosts) { mockLocalTracer }
-        fakeUrl = forgeUrlWithQueryParams(forge, forge.anElementFrom(fakeLocalHosts.keys))
+        fakeUrl = forgeUrl(forge, forge.anElementFrom(fakeLocalHosts.keys))
         fakeRequest = forgeRequest(forge)
         whenever(mockResolver.isFirstPartyUrl(HttpUrl.get(fakeUrl))).thenReturn(false)
         stubChain(mockChain, statusCode)
@@ -522,13 +444,15 @@ internal open class TracingInterceptorTest {
     ) {
         // Given
         whenever(mockTraceSampler.sample()).thenReturn(false)
-        fakeLocalHosts =
-            forge.aMap { forge.aStringMatching(HOSTNAME_PATTERN) to setOf(TracingHeaderType.B3) }
+        fakeLocalHosts = forge.aMap {
+            forge.aStringMatching(TracingInterceptorTest.HOSTNAME_PATTERN) to setOf(
+                TracingHeaderType.B3
+            )
+        }
         testedInterceptor = instantiateTestedInterceptor(fakeLocalHosts) { mockLocalTracer }
-        fakeUrl = forgeUrlWithQueryParams(forge, forge.anElementFrom(fakeLocalHosts.keys))
+        fakeUrl = forgeUrl(forge, forge.anElementFrom(fakeLocalHosts.keys))
         fakeRequest = forgeRequest(forge)
         whenever(mockResolver.isFirstPartyUrl(HttpUrl.get(fakeUrl))).thenReturn(false)
-
         stubChain(mockChain, statusCode)
 
         // When
@@ -550,54 +474,15 @@ internal open class TracingInterceptorTest {
     ) {
         // Given
         whenever(mockTraceSampler.sample()).thenReturn(false)
-        fakeLocalHosts =
-            forge.aMap { forge.aStringMatching(HOSTNAME_PATTERN) to setOf(TracingHeaderType.TRACECONTEXT) }
-        testedInterceptor = instantiateTestedInterceptor(fakeLocalHosts) { mockLocalTracer }
-
-        fakeUrl = forgeUrlWithQueryParams(forge, forge.anElementFrom(fakeLocalHosts.keys))
-        fakeRequest = forgeRequest(forge)
-        whenever(mockResolver.isFirstPartyUrl(HttpUrl.get(fakeUrl))).thenReturn(false)
-
-        stubChain(mockChain, statusCode)
-
-        // When
-        val response = testedInterceptor.intercept(mockChain)
-
-        // Then
-        assertThat(response).isSameAs(fakeResponse)
-        argumentCaptor<Request> {
-            verify(mockChain).proceed(capture())
-            assertThat(lastValue.header(TracingInterceptor.W3C_TRACEPARENT_KEY))
-                .isEqualTo(
-                    "00-%s-%s-00".format(
-                        mockSpan.context().toTraceId(),
-                        mockSpan.context().toSpanId()
-                    )
-                )
-        }
-    }
-
-    @Test
-    fun `𝕄 inject all non-tracing headers 𝕎 intercept() {local known host + not sampled}`(
-        @IntForgery(min = 200, max = 600) statusCode: Int,
-        forge: Forge
-    ) {
-        // Given
-        whenever(mockTraceSampler.sample()).thenReturn(false)
         fakeLocalHosts = forge.aMap {
-            forge.aStringMatching(HOSTNAME_PATTERN) to setOf(
-                TracingHeaderType.DATADOG,
-                TracingHeaderType.B3,
-                TracingHeaderType.B3MULTI,
+            forge.aStringMatching(TracingInterceptorTest.HOSTNAME_PATTERN) to setOf(
                 TracingHeaderType.TRACECONTEXT
             )
         }
         testedInterceptor = instantiateTestedInterceptor(fakeLocalHosts) { mockLocalTracer }
-
-        fakeUrl = forgeUrlWithQueryParams(forge, forge.anElementFrom(fakeLocalHosts.keys))
+        fakeUrl = forgeUrl(forge, forge.anElementFrom(fakeLocalHosts.keys))
         fakeRequest = forgeRequest(forge)
         whenever(mockResolver.isFirstPartyUrl(HttpUrl.get(fakeUrl))).thenReturn(false)
-
         stubChain(mockChain, statusCode)
 
         // When
@@ -607,17 +492,6 @@ internal open class TracingInterceptorTest {
         assertThat(response).isSameAs(fakeResponse)
         argumentCaptor<Request> {
             verify(mockChain).proceed(capture())
-            assertThat(lastValue.header(TracingInterceptor.DATADOG_SAMPLING_PRIORITY_HEADER))
-                .isEqualTo("0")
-            assertThat(lastValue.header(TracingInterceptor.DATADOG_TRACE_ID_HEADER)).isNull()
-            assertThat(lastValue.header(TracingInterceptor.DATADOG_SPAN_ID_HEADER)).isNull()
-            assertThat(lastValue.header(TracingInterceptor.DATADOG_ORIGIN_HEADER)).isNull()
-            assertThat(lastValue.header(TracingInterceptor.B3M_SAMPLING_PRIORITY_KEY))
-                .isEqualTo("0")
-            assertThat(lastValue.header(TracingInterceptor.B3M_SPAN_ID_KEY)).isNull()
-            assertThat(lastValue.header(TracingInterceptor.B3M_TRACE_ID_KEY)).isNull()
-            assertThat(lastValue.header(TracingInterceptor.B3_HEADER_KEY))
-                .isEqualTo("0")
             assertThat(lastValue.header(TracingInterceptor.W3C_TRACEPARENT_KEY))
                 .isEqualTo(
                     "00-%s-%s-00".format(
@@ -641,6 +515,7 @@ internal open class TracingInterceptorTest {
         whenever(mockSpanBuilder.asChildOf(parentSpanContext)) doReturn mockSpanBuilder
         fakeRequest = forgeRequest(forge) { it.tag(Span::class.java, parentSpan) }
         whenever(mockResolver.isFirstPartyUrl(HttpUrl.get(fakeUrl))).thenReturn(true)
+        doAnswer { true }.whenever(mockResolver).isFirstPartyUrl(HttpUrl.get(fakeUrl))
         stubChain(mockChain, statusCode)
         doAnswer { invocation ->
             val carrier = invocation.arguments[2] as TextMapInject
@@ -652,94 +527,24 @@ internal open class TracingInterceptorTest {
         assertThat(response).isSameAs(fakeResponse)
         argumentCaptor<Request> {
             verify(mockChain).proceed(capture())
-            assertThat(firstValue.headers(key)).containsOnly(value)
+            assertThat(lastValue.header(key)).isEqualTo(value)
         }
         verify(mockSpanBuilder).asChildOf(parentSpanContext)
-        verify(mockSpanBuilder).withOrigin(getExpectedOrigin())
-    }
-
-    @Test
-    fun `𝕄 replace existing tracing header 𝕎 intercept() {global known host}`(
-        @StringForgery key: String,
-        @StringForgery(type = StringForgeryType.HEXADECIMAL) value: String,
-        @StringForgery(type = StringForgeryType.ALPHABETICAL) previousValue: String,
-        @IntForgery(min = 200, max = 600) statusCode: Int
-    ) {
-        fakeRequest = fakeRequest.newBuilder().addHeader(key, previousValue).build()
-        stubChain(mockChain, statusCode)
-        whenever(mockResolver.isFirstPartyUrl(HttpUrl.get(fakeUrl))).thenReturn(true)
-        doAnswer { invocation ->
-            val carrier = invocation.arguments[2] as TextMapInject
-            carrier.put(key, value)
-        }.whenever(mockTracer).inject<TextMapInject>(any(), any(), any())
-
-        val response = testedInterceptor.intercept(mockChain)
-
-        assertThat(response).isSameAs(fakeResponse)
-        argumentCaptor<Request> {
-            verify(mockChain).proceed(capture())
-            assertThat(firstValue.headers(key)).containsOnly(value)
-        }
-    }
-
-    @Test
-    fun `𝕄 replace existing tracing header 𝕎 intercept() {local known host}`(
-        @StringForgery key: String,
-        @StringForgery(type = StringForgeryType.ALPHA_NUMERICAL) value: String,
-        @StringForgery(type = StringForgeryType.ALPHABETICAL) previousValue: String,
-        @IntForgery(min = 200, max = 600) statusCode: Int,
-        forge: Forge
-    ) {
-        fakeUrl = forgeUrlWithQueryParams(forge, forge.anElementFrom(fakeLocalHosts.keys))
-        fakeRequest = forgeRequest(forge).newBuilder().addHeader(key, previousValue).build()
-        whenever(mockResolver.isFirstPartyUrl(HttpUrl.get(fakeUrl))).thenReturn(false)
-        stubChain(mockChain, statusCode)
-        doAnswer { invocation ->
-            val carrier = invocation.arguments[2] as TextMapInject
-            carrier.put(key, value)
-        }.whenever(mockTracer).inject<TextMapInject>(any(), any(), any())
-
-        val response = testedInterceptor.intercept(mockChain)
-
-        assertThat(response).isSameAs(fakeResponse)
-        argumentCaptor<Request> {
-            verify(mockChain).proceed(capture())
-            assertThat(firstValue.headers(key)).containsOnly(value)
-        }
-    }
-
-    @Test
-    fun `𝕄 ignore inject exception 𝕎 intercept() {IllegalStateException}`(
-        @IntForgery(min = 200, max = 600) statusCode: Int,
-        @StringForgery message: String,
-        forge: Forge
-    ) {
-        fakeUrl = forgeUrlWithQueryParams(forge, forge.anElementFrom(fakeLocalHosts.keys))
-        fakeRequest = forgeRequest(forge)
-        whenever(mockResolver.isFirstPartyUrl(HttpUrl.get(fakeUrl))).thenReturn(false)
-        stubChain(mockChain, statusCode)
-        doThrow(IllegalStateException(message)).whenever(mockTracer)
-            .inject<TextMapInject>(any(), any(), any())
-
-        val response = testedInterceptor.intercept(mockChain)
-
-        assertThat(response).isSameAs(fakeResponse)
-        argumentCaptor<Request> {
-            verify(mockChain).proceed(capture())
-            assertThat(firstValue.headers().size()).isZero()
-        }
     }
 
     @Test
     fun `𝕄 update header with parent context 𝕎 intercept() for request with tracing headers`(
         @StringForgery key: String,
         @StringForgery(type = StringForgeryType.ALPHA_NUMERICAL) value: String,
-        @IntForgery(min = 200, max = 300) statusCode: Int
+        @IntForgery(min = 200, max = 300) statusCode: Int,
+        forge: Forge
     ) {
         val parentSpanContext: SpanContext = mock()
         whenever(mockTracer.extract<TextMapExtract>(any(), any())) doReturn parentSpanContext
         whenever(mockSpanBuilder.asChildOf(any<SpanContext>())) doReturn mockSpanBuilder
         whenever(mockResolver.isFirstPartyUrl(HttpUrl.get(fakeUrl))).thenReturn(true)
+        fakeRequest = forgeRequest(forge)
+        doAnswer { true }.whenever(mockResolver).isFirstPartyUrl(HttpUrl.get(fakeUrl))
         stubChain(mockChain, statusCode)
         doAnswer { invocation ->
             val carrier = invocation.arguments[2] as TextMapInject
@@ -751,10 +556,9 @@ internal open class TracingInterceptorTest {
         assertThat(response).isSameAs(fakeResponse)
         argumentCaptor<Request> {
             verify(mockChain).proceed(capture())
-            assertThat(firstValue.headers(key)).containsOnly(value)
+            assertThat(lastValue.header(key)).isEqualTo(value)
         }
         verify(mockSpanBuilder).asChildOf(parentSpanContext)
-        verify(mockSpanBuilder).withOrigin(getExpectedOrigin())
     }
 
     @Test
@@ -835,7 +639,6 @@ internal open class TracingInterceptorTest {
     ) {
         // Given
         whenever(mockResolver.isFirstPartyUrl(HttpUrl.get(fakeUrl))).thenReturn(true)
-
         fakeRequest = forgeRequest(forge) {
             it.addHeader(
                 TracingInterceptor.B3_HEADER_KEY,
@@ -869,7 +672,6 @@ internal open class TracingInterceptorTest {
     ) {
         // Given
         whenever(mockResolver.isFirstPartyUrl(HttpUrl.get(fakeUrl))).thenReturn(true)
-
         fakeRequest = forgeRequest(forge) {
             it.addHeader(
                 TracingInterceptor.W3C_TRACEPARENT_KEY,
@@ -906,6 +708,7 @@ internal open class TracingInterceptorTest {
                 TracingHeaderType.DATADOG
             )
         )
+
         fakeRequest = forgeRequest(forge) {
             it.addHeader(
                 TracingInterceptor.DATADOG_SAMPLING_PRIORITY_HEADER,
@@ -944,6 +747,7 @@ internal open class TracingInterceptorTest {
                 TracingHeaderType.B3MULTI
             )
         )
+
         fakeRequest = forgeRequest(forge) {
             it.addHeader(
                 TracingInterceptor.B3M_SAMPLING_PRIORITY_KEY,
@@ -979,6 +783,7 @@ internal open class TracingInterceptorTest {
                 TracingHeaderType.B3
             )
         )
+
         fakeRequest = forgeRequest(forge) {
             it.addHeader(
                 TracingInterceptor.B3_HEADER_KEY,
@@ -1015,6 +820,7 @@ internal open class TracingInterceptorTest {
                 TracingHeaderType.TRACECONTEXT
             )
         )
+
         fakeRequest = forgeRequest(forge) {
             it.addHeader(
                 TracingInterceptor.W3C_TRACEPARENT_KEY,
@@ -1042,6 +848,22 @@ internal open class TracingInterceptorTest {
     }
 
     @Test
+    fun `𝕄 not create a span 𝕎 intercept() { not sampled }`(
+        @IntForgery(min = 200, max = 600) statusCode: Int
+    ) {
+        // Given
+        whenever(mockTraceSampler.sample()).thenReturn(false)
+        whenever(mockResolver.isFirstPartyUrl(HttpUrl.get(fakeUrl))).thenReturn(true)
+        stubChain(mockChain, statusCode)
+
+        // When
+        val response = testedInterceptor.intercept(mockChain)
+
+        // Then
+        assertThat(response).isSameAs(fakeResponse)
+    }
+
+    @Test
     fun `𝕄 create a span with info 𝕎 intercept() for successful request`(
         @IntForgery(min = 200, max = 300) statusCode: Int
     ) {
@@ -1050,35 +872,10 @@ internal open class TracingInterceptorTest {
 
         val response = testedInterceptor.intercept(mockChain)
 
-        verify(mockSpanBuilder).withOrigin(getExpectedOrigin())
-        verify(mockSpan as MutableSpan).resourceName = fakeBaseUrl.lowercase(Locale.US)
         verify(mockSpan).setTag("http.url", fakeUrl.lowercase(Locale.US))
         verify(mockSpan).setTag("http.method", fakeMethod)
         verify(mockSpan).setTag("http.status_code", statusCode)
-        verify(mockSpan).finish()
-        assertThat(response).isSameAs(fakeResponse)
-    }
-
-    @Test
-    fun `𝕄 create a span with info 𝕎 intercept() { resource url with no query paramaters }`(
-        @IntForgery(min = 200, max = 300) statusCode: Int,
-        forge: Forge
-    ) {
-        val fakeUrlWithoutQueryParams = forgeUrlWithoutQueryParams(forge)
-        fakeRequest = forgeRequest(forge, fakeUrlWithoutQueryParams)
-        whenever(mockResolver.isFirstPartyUrl(HttpUrl.get(fakeUrlWithoutQueryParams)))
-            .thenReturn(true)
-        stubChain(mockChain, statusCode)
-
-        val response = testedInterceptor.intercept(mockChain)
-
-        verify(mockSpanBuilder).withOrigin(getExpectedOrigin())
-        verify(mockSpan as MutableSpan).resourceName =
-            fakeUrlWithoutQueryParams.lowercase(Locale.US)
-        verify(mockSpan).setTag("http.url", fakeUrlWithoutQueryParams.lowercase(Locale.US))
-        verify(mockSpan).setTag("http.method", fakeMethod)
-        verify(mockSpan).setTag("http.status_code", statusCode)
-        verify(mockSpan).finish()
+        verify(mockSpan, never()).finish()
         assertThat(response).isSameAs(fakeResponse)
     }
 
@@ -1091,13 +888,10 @@ internal open class TracingInterceptorTest {
 
         val response = testedInterceptor.intercept(mockChain)
 
-        verify(mockSpanBuilder).withOrigin(getExpectedOrigin())
-        verify(mockSpan as MutableSpan).resourceName = fakeBaseUrl.lowercase(Locale.US)
         verify(mockSpan).setTag("http.url", fakeUrl.lowercase(Locale.US))
         verify(mockSpan).setTag("http.method", fakeMethod)
         verify(mockSpan).setTag("http.status_code", statusCode)
-        verify(mockSpan as MutableSpan).setError(true)
-        verify(mockSpan).finish()
+        verify(mockSpan, never()).finish()
         assertThat(response).isSameAs(fakeResponse)
     }
 
@@ -1110,13 +904,10 @@ internal open class TracingInterceptorTest {
 
         val response = testedInterceptor.intercept(mockChain)
 
-        verify(mockSpanBuilder).withOrigin(getExpectedOrigin())
-        verify(mockSpan as MutableSpan).resourceName = fakeBaseUrl.lowercase(Locale.US)
         verify(mockSpan).setTag("http.url", fakeUrl.lowercase(Locale.US))
         verify(mockSpan).setTag("http.method", fakeMethod)
         verify(mockSpan).setTag("http.status_code", statusCode)
-        verify(mockSpan as MutableSpan, never()).setError(true)
-        verify(mockSpan).finish()
+        verify(mockSpan, never()).finish()
         assertThat(response).isSameAs(fakeResponse)
     }
 
@@ -1127,13 +918,10 @@ internal open class TracingInterceptorTest {
 
         val response = testedInterceptor.intercept(mockChain)
 
-        verify(mockSpanBuilder).withOrigin(getExpectedOrigin())
         verify(mockSpan).setTag("http.url", fakeUrl.lowercase(Locale.US))
         verify(mockSpan).setTag("http.method", fakeMethod)
         verify(mockSpan).setTag("http.status_code", 404)
-        verify(mockSpan as MutableSpan).setError(true)
-        verify(mockSpan as MutableSpan).setResourceName(TracingInterceptor.RESOURCE_NAME_404)
-        verify(mockSpan).finish()
+        verify(mockSpan, never()).finish()
         assertThat(response).isSameAs(fakeResponse)
     }
 
@@ -1149,14 +937,12 @@ internal open class TracingInterceptorTest {
             testedInterceptor.intercept(mockChain)
         }
 
-        verify(mockSpanBuilder).withOrigin(getExpectedOrigin())
         verify(mockSpan).setTag("http.url", fakeUrl.lowercase(Locale.US))
         verify(mockSpan).setTag("http.method", fakeMethod)
         verify(mockSpan).setTag("error.type", throwable.javaClass.canonicalName)
         verify(mockSpan).setTag("error.msg", throwable.message)
         verify(mockSpan).setTag("error.stack", throwable.loggableStackTrace())
-        verify(mockSpan as MutableSpan).resourceName = fakeBaseUrl.lowercase(Locale.US)
-        verify(mockSpan).finish()
+        verify(mockSpan, never()).finish()
     }
 
     @Test
@@ -1165,7 +951,6 @@ internal open class TracingInterceptorTest {
     ) {
         GlobalTracer::class.java.setStaticValue("isRegistered", false)
         Datadog.globalSdkCore = NoOpSdkCore()
-        whenever(mockResolver.isFirstPartyUrl(HttpUrl.get(fakeUrl))).thenReturn(true)
         stubChain(mockChain, statusCode)
 
         testedInterceptor.intercept(mockChain)
@@ -1184,13 +969,12 @@ internal open class TracingInterceptorTest {
     fun `𝕄 create a span with automatic tracer 𝕎 intercept() if no tracer registered`(
         @IntForgery(min = 200, max = 300) statusCode: Int
     ) {
-        val localSpanBuilder: DDTracer.DDSpanBuilder = mock()
-        val localSpan: Span = mock(extraInterfaces = arrayOf(MutableSpan::class))
+        val localSpanBuilder: Tracer.SpanBuilder = mock()
+        val localSpan: Span = mock()
         GlobalTracer::class.java.setStaticValue("isRegistered", false)
         whenever(mockResolver.isFirstPartyUrl(HttpUrl.get(fakeUrl))).thenReturn(true)
         stubChain(mockChain, statusCode)
         whenever(localSpanBuilder.asChildOf(null as SpanContext?)) doReturn localSpanBuilder
-        whenever(localSpanBuilder.withOrigin(getExpectedOrigin())) doReturn localSpanBuilder
         whenever(localSpanBuilder.start()) doReturn localSpan
         whenever(localSpan.context()) doReturn mockSpanContext
         whenever(mockSpanContext.toSpanId()) doReturn fakeSpanId
@@ -1199,12 +983,10 @@ internal open class TracingInterceptorTest {
 
         val response = testedInterceptor.intercept(mockChain)
 
-        verify(localSpanBuilder).withOrigin(getExpectedOrigin())
         verify(localSpan).setTag("http.url", fakeUrl.lowercase(Locale.US))
         verify(localSpan).setTag("http.method", fakeMethod)
         verify(localSpan).setTag("http.status_code", statusCode)
-        verify(localSpan as MutableSpan).resourceName = fakeBaseUrl.lowercase(Locale.US)
-        verify(localSpan).finish()
+        verify(localSpan, never()).finish()
         assertThat(response).isSameAs(fakeResponse)
         verify(logger.mockInternalLogger)
             .log(
@@ -1218,13 +1000,12 @@ internal open class TracingInterceptorTest {
     fun `𝕄 drop automatic tracer 𝕎 intercept() and global tracer registered`(
         @IntForgery(min = 200, max = 300) statusCode: Int
     ) {
-        val localSpanBuilder: DDTracer.DDSpanBuilder = mock()
-        val localSpan: Span = mock(extraInterfaces = arrayOf(MutableSpan::class))
+        val localSpanBuilder: Tracer.SpanBuilder = mock()
+        val localSpan: Span = mock()
         GlobalTracer::class.java.setStaticValue("isRegistered", false)
         whenever(mockResolver.isFirstPartyUrl(HttpUrl.get(fakeUrl))).thenReturn(true)
         stubChain(mockChain, statusCode)
         whenever(localSpanBuilder.asChildOf(null as SpanContext?)) doReturn localSpanBuilder
-        whenever(localSpanBuilder.withOrigin(getExpectedOrigin())) doReturn localSpanBuilder
         whenever(localSpanBuilder.start()) doReturn localSpan
         whenever(localSpan.context()) doReturn mockSpanContext
         whenever(mockSpanContext.toSpanId()) doReturn fakeSpanId
@@ -1238,18 +1019,14 @@ internal open class TracingInterceptorTest {
         val response2 = testedInterceptor.intercept(mockChain)
         val expectedResponse2 = fakeResponse
 
-        verify(localSpanBuilder).withOrigin(getExpectedOrigin())
         verify(localSpan).setTag("http.url", fakeUrl.lowercase(Locale.US))
         verify(localSpan).setTag("http.method", fakeMethod)
         verify(localSpan).setTag("http.status_code", statusCode)
-        verify(localSpan as MutableSpan).resourceName = fakeBaseUrl.lowercase(Locale.US)
-        verify(localSpan).finish()
-        verify(mockSpanBuilder).withOrigin(getExpectedOrigin())
+        verify(localSpan, never()).finish()
         verify(mockSpan).setTag("http.url", fakeUrl.lowercase(Locale.US))
         verify(mockSpan).setTag("http.method", fakeMethod)
         verify(mockSpan).setTag("http.status_code", statusCode)
-        verify(mockSpan as MutableSpan).resourceName = fakeBaseUrl.lowercase(Locale.US)
-        verify(mockSpan).finish()
+        verify(mockSpan, never()).finish()
         assertThat(response1).isSameAs(expectedResponse1)
         assertThat(response2).isSameAs(expectedResponse2)
         verify(logger.mockInternalLogger)
@@ -1278,13 +1055,11 @@ internal open class TracingInterceptorTest {
 
         val response = testedInterceptor.intercept(mockChain)
 
-        verify(mockSpanBuilder).withOrigin(getExpectedOrigin())
         verify(mockSpan).setTag("http.url", fakeUrl.lowercase(Locale.US))
         verify(mockSpan).setTag("http.method", fakeMethod)
         verify(mockSpan).setTag("http.status_code", statusCode)
-        verify(mockSpan as MutableSpan).resourceName = fakeBaseUrl.lowercase(Locale.US)
         verify(mockSpan).setTag(tagKey, tagValue)
-        verify(mockSpan).finish()
+        verify(mockSpan, never()).finish()
         assertThat(response).isSameAs(fakeResponse)
     }
 
@@ -1306,12 +1081,11 @@ internal open class TracingInterceptorTest {
 
         val response = testedInterceptor.intercept(mockChain)
 
-        verify(mockSpanBuilder).withOrigin(getExpectedOrigin())
         verify(mockSpan).setTag("http.url", fakeUrl.lowercase(Locale.US))
         verify(mockSpan).setTag("http.method", fakeMethod)
         verify(mockSpan).setTag("http.status_code", statusCode)
         verify(mockSpan).setTag(tagKey, tagValue)
-        verify(mockSpan).finish()
+        verify(mockSpan, never()).finish()
         assertThat(response).isSameAs(fakeResponse)
     }
 
@@ -1353,14 +1127,13 @@ internal open class TracingInterceptorTest {
             testedInterceptor.intercept(mockChain)
         }
 
-        verify(mockSpanBuilder).withOrigin(getExpectedOrigin())
         verify(mockSpan).setTag("http.url", fakeUrl.lowercase(Locale.US))
         verify(mockSpan).setTag("http.method", fakeMethod)
         verify(mockSpan).setTag("error.type", throwable.javaClass.canonicalName)
         verify(mockSpan).setTag("error.msg", throwable.message)
         verify(mockSpan).setTag("error.stack", throwable.loggableStackTrace())
         verify(mockSpan).setTag(tagKey, tagValue)
-        verify(mockSpan).finish()
+        verify(mockSpan, never()).finish()
     }
 
     @Test
@@ -1383,20 +1156,28 @@ internal open class TracingInterceptorTest {
     @Test
     fun `M do not update the hostDetector W host list provided`(forge: Forge) {
         // GIVEN
-        val localHosts =
-            forge.aMap { aStringMatching(HOSTNAME_PATTERN) to setOf(TracingHeaderType.DATADOG) }
+        val localHosts = forge.aList { forge.aStringMatching(HOSTNAME_PATTERN) }
 
         // WHEN
-        testedInterceptor = instantiateTestedInterceptor(localHosts) { mockLocalTracer }
+        testedInterceptor =
+            instantiateTestedInterceptor(
+                localHosts.associateWith {
+                    setOf(TracingHeaderType.DATADOG)
+                }
+            ) {
+                mockLocalTracer
+            }
 
         // THEN
-        verify(mockResolver, never()).addKnownHostsWithHeaderTypes(localHosts)
+        verify(mockResolver, never()).addKnownHosts(localHosts)
     }
 
     @Test
     fun `𝕄 do nothing 𝕎 intercept() for request with unknown host`(
-        @IntForgery(min = 200, max = 300) statusCode: Int
+        @IntForgery(min = 200, max = 300) statusCode: Int,
+        forge: Forge
     ) {
+        fakeRequest = forgeRequest(forge)
         whenever(mockResolver.isFirstPartyUrl(HttpUrl.get(fakeUrl))).thenReturn(false)
         stubChain(mockChain, statusCode)
 
@@ -1408,11 +1189,9 @@ internal open class TracingInterceptorTest {
 
     @Test
     fun `𝕄 warn 𝕎 init() with no known host`() {
-        // GIVEN
-        whenever(mockResolver.isEmpty()).thenReturn(true)
+        whenever(mockResolver.isEmpty()) doReturn true
 
-        // WHEN
-        testedInterceptor = instantiateTestedInterceptor { mockLocalTracer }
+        testedInterceptor = instantiateTestedInterceptor(emptyMap()) { mockLocalTracer }
 
         verifyZeroInteractions(mockTracer, mockLocalTracer)
         verify(logger.mockInternalLogger)
@@ -1429,12 +1208,14 @@ internal open class TracingInterceptorTest {
     ) {
         // Given
         var called = 0
-        testedInterceptor = instantiateTestedInterceptor {
+        val tracedHosts =
+            listOf(fakeHostIp, fakeHostName).associateWith { setOf(TracingHeaderType.DATADOG) }
+        whenever(mockResolver.isFirstPartyUrl(HttpUrl.get(fakeUrl))).thenReturn(true)
+        testedInterceptor = instantiateTestedInterceptor(tracedHosts) {
             called++
             mockLocalTracer
         }
         GlobalTracer::class.java.setStaticValue("isRegistered", false)
-        whenever(mockResolver.isFirstPartyUrl(HttpUrl.get(fakeUrl))).thenReturn(true)
         stubChain(mockChain, statusCode)
 
         // need this setup, otherwise #intercept actually throws NPE, which pollutes the log
@@ -1459,41 +1240,39 @@ internal open class TracingInterceptorTest {
         }.start()
 
         // Then
-        countDownLatch.await(5, TimeUnit.SECONDS)
+        countDownLatch.await(1, TimeUnit.SECONDS)
         verify(mockLocalTracer, times(2)).buildSpan(TracingInterceptor.SPAN_NAME)
         assertThat(called).isEqualTo(1)
     }
 
     // region Internal
 
-    internal fun stubChain(chain: Interceptor.Chain, statusCode: Int) {
-        fakeResponse = forgeResponse(statusCode)
+    private fun stubChain(chain: Interceptor.Chain, statusCode: Int) {
+        return stubChain(chain) { forgeResponse(statusCode) }
+    }
+
+    private fun stubChain(
+        chain: Interceptor.Chain,
+        responseBuilder: () -> Response
+    ) {
+        fakeResponse = responseBuilder()
 
         whenever(chain.request()) doReturn fakeRequest
         whenever(chain.proceed(any())) doReturn fakeResponse
     }
 
-    private fun forgeUrlWithoutQueryParams(forge: Forge, knownHost: String? = null): String {
+    private fun forgeUrl(forge: Forge, knownHost: String? = null): String {
         val protocol = forge.anElementFrom("http", "https")
-        val host = knownHost ?: forge.aStringMatching(HOSTNAME_PATTERN)
+        val host = knownHost ?: forge.aStringMatching(TracingInterceptorTest.HOSTNAME_PATTERN)
         val path = forge.anAlphaNumericalString()
         return "$protocol://$host/$path"
     }
 
-    private fun forgeUrlWithQueryParams(forge: Forge, knownHost: String? = null): String {
-        fakeBaseUrl = forgeUrlWithoutQueryParams(forge, knownHost)
-        val fakeQueryParams = forge.aList(forge.anInt(min = 0, max = 5)) {
-            "${forge.anAlphabeticalString()}=${forge.anAlphabeticalString()}"
-        }.joinToString("&")
-        return "$fakeBaseUrl?$fakeQueryParams"
-    }
-
     private fun forgeRequest(
         forge: Forge,
-        url: String = fakeUrl,
         configure: (Request.Builder) -> Unit = {}
     ): Request {
-        val builder = Request.Builder().url(url)
+        val builder = Request.Builder().url(fakeUrl)
         if (forge.aBool()) {
             fakeMethod = "POST"
             fakeBody = forge.anAlphabeticalString()
@@ -1515,8 +1294,10 @@ internal open class TracingInterceptorTest {
             .protocol(Protocol.HTTP_2)
             .code(statusCode)
             .message("HTTP $statusCode")
-            .header(TracingInterceptor.HEADER_CT, fakeMediaType?.type().orEmpty())
             .body(ResponseBody.create(fakeMediaType, fakeResponseBody))
+        if (fakeMediaType != null) {
+            builder.header(TracingInterceptor.HEADER_CT, fakeMediaType?.type().orEmpty())
+        }
         return builder.build()
     }
 
@@ -1526,12 +1307,18 @@ internal open class TracingInterceptorTest {
         const val HOSTNAME_PATTERN =
             "(([a-zA-Z0-9]|[a-zA-Z0-9][a-zA-Z0-9\\-]{1,4}[a-zA-Z0-9]{2,3})\\.)+" +
                 "([A-Za-z]|[A-Za-z][A-Za-z0-9-]{1,2}[A-Za-z0-9])"
+        const val IPV4_PATTERN =
+            "(([0-9]|[1-9][0-9]|1[0-9]){2}\\.|(2[0-4][0-9]|25[0-5])\\.){3}" +
+                "([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])"
+
+        val appContext = ApplicationContextTestConfiguration(Context::class.java)
+        val coreFeature = CoreFeatureTestConfiguration(appContext)
         val logger = InternalLoggerTestConfiguration()
 
         @TestConfigurationsProvider
         @JvmStatic
         fun getTestConfigurations(): List<TestConfiguration> {
-            return listOf(logger)
+            return listOf(logger, appContext, coreFeature)
         }
     }
 }
