+++ dd-sdk-android-okhttp/src/test/kotlin/com/datadog/android/okhttp/tracing/TracingInterceptorNotSendingSpanTest.kt
@@ -4,15 +4,15 @@
  * Copyright 2016-Present Datadog, Inc.
  */
 
-package com.datadog.android.tracing
+package com.datadog.android.okhttp.tracing
 
 import android.content.Context
 import android.util.Log
 import com.datadog.android.Datadog
-import com.datadog.android.core.configuration.Configuration
 import com.datadog.android.core.internal.net.FirstPartyHostHeaderTypeResolver
 import com.datadog.android.core.internal.utils.loggableStackTrace
 import com.datadog.android.core.sampling.Sampler
+import com.datadog.android.tracing.TracingHeaderType
 import com.datadog.android.utils.config.ApplicationContextTestConfiguration
 import com.datadog.android.utils.config.CoreFeatureTestConfiguration
 import com.datadog.android.utils.config.InternalLoggerTestConfiguration
@@ -20,6 +20,7 @@ import com.datadog.android.utils.forge.Configurator
 import com.datadog.android.v2.api.InternalLogger
 import com.datadog.android.v2.core.DatadogCore
 import com.datadog.android.v2.core.NoOpSdkCore
+import com.datadog.opentracing.DDSpanContext
 import com.datadog.opentracing.DDTracer
 import com.datadog.tools.unit.annotations.TestConfigurationsProvider
 import com.datadog.tools.unit.extensions.TestConfigurationExtension
@@ -75,10 +76,6 @@ import java.util.Locale
 import java.util.concurrent.CountDownLatch
 import java.util.concurrent.TimeUnit
 
-/**
- * This test is a duplicate of [TracingInterceptorNotSendingSpanTest] but assuming the Tracer is
- * not our own implementation and therefore doesn't implement DD specific methods.
- */
 @Extensions(
     ExtendWith(MockitoExtension::class),
     ExtendWith(ForgeExtension::class),
@@ -86,7 +83,7 @@ import java.util.concurrent.TimeUnit
 )
 @MockitoSettings(strictness = Strictness.LENIENT)
 @ForgeConfiguration(Configurator::class)
-internal open class TracingInterceptorNonDdTracerNotSendingSpanTest {
+internal open class TracingInterceptorNotSendingSpanTest {
 
     lateinit var testedInterceptor: TracingInterceptor
 
@@ -99,12 +96,12 @@ internal open class TracingInterceptorNonDdTracerNotSendingSpanTest {
     lateinit var mockLocalTracer: Tracer
 
     @Mock
-    lateinit var mockSpanBuilder: Tracer.SpanBuilder
+    lateinit var mockSpanBuilder: DDTracer.DDSpanBuilder
 
     @Mock
-    lateinit var mockSpanContext: SpanContext
+    lateinit var mockSpanContext: DDSpanContext
 
-    @Mock
+    @Mock(extraInterfaces = [MutableSpan::class])
     lateinit var mockSpan: Span
 
     @Mock
@@ -137,9 +134,6 @@ internal open class TracingInterceptorNonDdTracerNotSendingSpanTest {
     lateinit var fakeResponseBody: String
 
     lateinit var fakeUrl: String
-
-    @Forgery
-    lateinit var fakeConfig: Configuration.Feature.Tracing
 
     lateinit var fakeRequest: Request
     lateinit var fakeResponse: Response
@@ -183,9 +177,8 @@ internal open class TracingInterceptorNonDdTracerNotSendingSpanTest {
         }
         fakeUrl = forgeUrl(forge)
         fakeRequest = forgeRequest(forge)
-        val mockCore = mock<DatadogCore>()
-        whenever(mockCore.tracingFeature) doReturn mock()
-        Datadog.globalSdkCore = mockCore
+        Datadog.globalSdkCore = mock<DatadogCore>()
+        whenever((Datadog.globalSdkCore as DatadogCore).tracingFeature) doReturn mock()
         doAnswer { false }.whenever(mockResolver).isFirstPartyUrl(any<HttpUrl>())
         doAnswer { true }.whenever(mockResolver).isFirstPartyUrl(HttpUrl.get(fakeUrl))
 
@@ -218,13 +211,30 @@ internal open class TracingInterceptorNonDdTracerNotSendingSpanTest {
         }
     }
 
+    open fun getExpectedOrigin(): String {
+        return fakeOrigin
+    }
+
     @Test
     fun `𝕄 inject tracing header 𝕎 intercept() {global known host}`(
         @StringForgery key: String,
         @StringForgery(type = StringForgeryType.ALPHA_NUMERICAL) value: String,
-        @IntForgery(min = 200, max = 600) statusCode: Int
+        @IntForgery(min = 200, max = 600) statusCode: Int,
+        forge: Forge
     ) {
         whenever(mockResolver.isFirstPartyUrl(HttpUrl.get(fakeUrl))).thenReturn(true)
+        whenever(mockResolver.headerTypesForUrl(HttpUrl.get(fakeUrl))).thenReturn(
+            setOf(
+                forge.anElementFrom(
+                    setOf(
+                        TracingHeaderType.DATADOG,
+                        TracingHeaderType.B3,
+                        TracingHeaderType.B3MULTI,
+                        TracingHeaderType.TRACECONTEXT
+                    )
+                )
+            )
+        )
         stubChain(mockChain, statusCode)
         doAnswer { invocation ->
             val carrier = invocation.arguments[2] as TextMapInject
@@ -269,7 +279,7 @@ internal open class TracingInterceptorNonDdTracerNotSendingSpanTest {
     }
 
     @Test
-    fun `𝕄 inject non-tracing b3multi header 𝕎 intercept() {global known host + not sampled}`(
+    fun `𝕄 inject non-tracing b3mult header 𝕎 intercept() {global known host + not sampled}`(
         @IntForgery(min = 200, max = 600) statusCode: Int
     ) {
         // Given
@@ -343,6 +353,51 @@ internal open class TracingInterceptorNonDdTracerNotSendingSpanTest {
         assertThat(response).isSameAs(fakeResponse)
         argumentCaptor<Request> {
             verify(mockChain).proceed(capture())
+            assertThat(lastValue.header(TracingInterceptor.W3C_TRACEPARENT_KEY))
+                .isEqualTo(
+                    "00-%s-%s-00".format(
+                        mockSpan.context().toTraceId(),
+                        mockSpan.context().toSpanId()
+                    )
+                )
+        }
+    }
+
+    @Test
+    fun `𝕄 inject all non-tracing header 𝕎 intercept() {global known host + not sampled}`(
+        @IntForgery(min = 200, max = 600) statusCode: Int
+    ) {
+        // Given
+        whenever(mockTraceSampler.sample()).thenReturn(false)
+        whenever(mockResolver.isFirstPartyUrl(HttpUrl.get(fakeUrl))).thenReturn(true)
+        whenever(mockResolver.headerTypesForUrl(HttpUrl.get(fakeUrl))).thenReturn(
+            setOf(
+                TracingHeaderType.DATADOG,
+                TracingHeaderType.B3,
+                TracingHeaderType.B3MULTI,
+                TracingHeaderType.TRACECONTEXT
+            )
+        )
+        stubChain(mockChain, statusCode)
+
+        // When
+        val response = testedInterceptor.intercept(mockChain)
+
+        // Then
+        assertThat(response).isSameAs(fakeResponse)
+        argumentCaptor<Request> {
+            verify(mockChain).proceed(capture())
+            assertThat(lastValue.header(TracingInterceptor.DATADOG_SAMPLING_PRIORITY_HEADER))
+                .isEqualTo("0")
+            assertThat(lastValue.header(TracingInterceptor.DATADOG_TRACE_ID_HEADER)).isNull()
+            assertThat(lastValue.header(TracingInterceptor.DATADOG_SPAN_ID_HEADER)).isNull()
+            assertThat(lastValue.header(TracingInterceptor.DATADOG_ORIGIN_HEADER)).isNull()
+            assertThat(lastValue.header(TracingInterceptor.B3M_SAMPLING_PRIORITY_KEY))
+                .isEqualTo("0")
+            assertThat(lastValue.header(TracingInterceptor.B3M_SPAN_ID_KEY)).isNull()
+            assertThat(lastValue.header(TracingInterceptor.B3M_TRACE_ID_KEY)).isNull()
+            assertThat(lastValue.header(TracingInterceptor.B3_HEADER_KEY))
+                .isEqualTo("0")
             assertThat(lastValue.header(TracingInterceptor.W3C_TRACEPARENT_KEY))
                 .isEqualTo(
                     "00-%s-%s-00".format(
@@ -529,6 +584,7 @@ internal open class TracingInterceptorNonDdTracerNotSendingSpanTest {
             assertThat(lastValue.header(key)).isEqualTo(value)
         }
         verify(mockSpanBuilder).asChildOf(parentSpanContext)
+        verify(mockSpanBuilder).withOrigin(getExpectedOrigin())
     }
 
     @Test
@@ -558,6 +614,7 @@ internal open class TracingInterceptorNonDdTracerNotSendingSpanTest {
             assertThat(lastValue.header(key)).isEqualTo(value)
         }
         verify(mockSpanBuilder).asChildOf(parentSpanContext)
+        verify(mockSpanBuilder).withOrigin(getExpectedOrigin())
     }
 
     @Test
@@ -707,7 +764,6 @@ internal open class TracingInterceptorNonDdTracerNotSendingSpanTest {
                 TracingHeaderType.DATADOG
             )
         )
-
         fakeRequest = forgeRequest(forge) {
             it.addHeader(
                 TracingInterceptor.DATADOG_SAMPLING_PRIORITY_HEADER,
@@ -746,7 +802,6 @@ internal open class TracingInterceptorNonDdTracerNotSendingSpanTest {
                 TracingHeaderType.B3MULTI
             )
         )
-
         fakeRequest = forgeRequest(forge) {
             it.addHeader(
                 TracingInterceptor.B3M_SAMPLING_PRIORITY_KEY,
@@ -782,7 +837,6 @@ internal open class TracingInterceptorNonDdTracerNotSendingSpanTest {
                 TracingHeaderType.B3
             )
         )
-
         fakeRequest = forgeRequest(forge) {
             it.addHeader(
                 TracingInterceptor.B3_HEADER_KEY,
@@ -819,7 +873,6 @@ internal open class TracingInterceptorNonDdTracerNotSendingSpanTest {
                 TracingHeaderType.TRACECONTEXT
             )
         )
-
         fakeRequest = forgeRequest(forge) {
             it.addHeader(
                 TracingInterceptor.W3C_TRACEPARENT_KEY,
@@ -847,34 +900,19 @@ internal open class TracingInterceptorNonDdTracerNotSendingSpanTest {
     }
 
     @Test
-    fun `𝕄 not create a span 𝕎 intercept() { not sampled }`(
-        @IntForgery(min = 200, max = 600) statusCode: Int
-    ) {
-        // Given
-        whenever(mockTraceSampler.sample()).thenReturn(false)
-        whenever(mockResolver.isFirstPartyUrl(HttpUrl.get(fakeUrl))).thenReturn(true)
-        stubChain(mockChain, statusCode)
-
-        // When
-        val response = testedInterceptor.intercept(mockChain)
-
-        // Then
-        assertThat(response).isSameAs(fakeResponse)
-    }
-
-    @Test
     fun `𝕄 create a span with info 𝕎 intercept() for successful request`(
         @IntForgery(min = 200, max = 300) statusCode: Int
     ) {
-        whenever(mockResolver.isFirstPartyUrl(HttpUrl.get(fakeUrl))).thenReturn(true)
         stubChain(mockChain, statusCode)
 
         val response = testedInterceptor.intercept(mockChain)
 
+        verify(mockSpanBuilder).withOrigin(getExpectedOrigin())
         verify(mockSpan).setTag("http.url", fakeUrl.lowercase(Locale.US))
         verify(mockSpan).setTag("http.method", fakeMethod)
         verify(mockSpan).setTag("http.status_code", statusCode)
         verify(mockSpan, never()).finish()
+        verify(mockSpan as MutableSpan).drop()
         assertThat(response).isSameAs(fakeResponse)
     }
 
@@ -887,10 +925,14 @@ internal open class TracingInterceptorNonDdTracerNotSendingSpanTest {
 
         val response = testedInterceptor.intercept(mockChain)
 
+        verify(mockSpanBuilder).withOrigin(getExpectedOrigin())
+        verify(mockSpan as MutableSpan).setResourceName(fakeUrl.lowercase(Locale.US))
         verify(mockSpan).setTag("http.url", fakeUrl.lowercase(Locale.US))
         verify(mockSpan).setTag("http.method", fakeMethod)
         verify(mockSpan).setTag("http.status_code", statusCode)
+        verify(mockSpan as MutableSpan).setError(true)
         verify(mockSpan, never()).finish()
+        verify(mockSpan as MutableSpan).drop()
         assertThat(response).isSameAs(fakeResponse)
     }
 
@@ -903,10 +945,14 @@ internal open class TracingInterceptorNonDdTracerNotSendingSpanTest {
 
         val response = testedInterceptor.intercept(mockChain)
 
+        verify(mockSpanBuilder).withOrigin(getExpectedOrigin())
+        verify(mockSpan as MutableSpan).setResourceName(fakeUrl.lowercase(Locale.US))
         verify(mockSpan).setTag("http.url", fakeUrl.lowercase(Locale.US))
         verify(mockSpan).setTag("http.method", fakeMethod)
         verify(mockSpan).setTag("http.status_code", statusCode)
+        verify(mockSpan as MutableSpan, never()).setError(true)
         verify(mockSpan, never()).finish()
+        verify(mockSpan as MutableSpan).drop()
         assertThat(response).isSameAs(fakeResponse)
     }
 
@@ -917,10 +963,14 @@ internal open class TracingInterceptorNonDdTracerNotSendingSpanTest {
 
         val response = testedInterceptor.intercept(mockChain)
 
+        verify(mockSpanBuilder).withOrigin(getExpectedOrigin())
         verify(mockSpan).setTag("http.url", fakeUrl.lowercase(Locale.US))
         verify(mockSpan).setTag("http.method", fakeMethod)
         verify(mockSpan).setTag("http.status_code", 404)
+        verify(mockSpan as MutableSpan).setError(true)
+        verify(mockSpan as MutableSpan).setResourceName(TracingInterceptor.RESOURCE_NAME_404)
         verify(mockSpan, never()).finish()
+        verify(mockSpan as MutableSpan).drop()
         assertThat(response).isSameAs(fakeResponse)
     }
 
@@ -936,12 +986,14 @@ internal open class TracingInterceptorNonDdTracerNotSendingSpanTest {
             testedInterceptor.intercept(mockChain)
         }
 
+        verify(mockSpanBuilder).withOrigin(getExpectedOrigin())
         verify(mockSpan).setTag("http.url", fakeUrl.lowercase(Locale.US))
         verify(mockSpan).setTag("http.method", fakeMethod)
         verify(mockSpan).setTag("error.type", throwable.javaClass.canonicalName)
         verify(mockSpan).setTag("error.msg", throwable.message)
         verify(mockSpan).setTag("error.stack", throwable.loggableStackTrace())
         verify(mockSpan, never()).finish()
+        verify(mockSpan as MutableSpan).drop()
     }
 
     @Test
@@ -968,8 +1020,8 @@ internal open class TracingInterceptorNonDdTracerNotSendingSpanTest {
     fun `𝕄 create a span with automatic tracer 𝕎 intercept() if no tracer registered`(
         @IntForgery(min = 200, max = 300) statusCode: Int
     ) {
-        val localSpanBuilder: Tracer.SpanBuilder = mock()
-        val localSpan: Span = mock()
+        val localSpanBuilder: DDTracer.DDSpanBuilder = mock()
+        val localSpan: Span = mock(extraInterfaces = arrayOf(MutableSpan::class))
         GlobalTracer::class.java.setStaticValue("isRegistered", false)
         whenever(mockResolver.isFirstPartyUrl(HttpUrl.get(fakeUrl))).thenReturn(true)
         stubChain(mockChain, statusCode)
@@ -982,10 +1034,12 @@ internal open class TracingInterceptorNonDdTracerNotSendingSpanTest {
 
         val response = testedInterceptor.intercept(mockChain)
 
+        verify(localSpanBuilder).withOrigin(getExpectedOrigin())
         verify(localSpan).setTag("http.url", fakeUrl.lowercase(Locale.US))
         verify(localSpan).setTag("http.method", fakeMethod)
         verify(localSpan).setTag("http.status_code", statusCode)
         verify(localSpan, never()).finish()
+        verify(localSpan as MutableSpan).drop()
         assertThat(response).isSameAs(fakeResponse)
         verify(logger.mockInternalLogger)
             .log(
@@ -999,8 +1053,8 @@ internal open class TracingInterceptorNonDdTracerNotSendingSpanTest {
     fun `𝕄 drop automatic tracer 𝕎 intercept() and global tracer registered`(
         @IntForgery(min = 200, max = 300) statusCode: Int
     ) {
-        val localSpanBuilder: Tracer.SpanBuilder = mock()
-        val localSpan: Span = mock()
+        val localSpanBuilder: DDTracer.DDSpanBuilder = mock()
+        val localSpan: Span = mock(extraInterfaces = arrayOf(MutableSpan::class))
         GlobalTracer::class.java.setStaticValue("isRegistered", false)
         whenever(mockResolver.isFirstPartyUrl(HttpUrl.get(fakeUrl))).thenReturn(true)
         stubChain(mockChain, statusCode)
@@ -1018,14 +1072,18 @@ internal open class TracingInterceptorNonDdTracerNotSendingSpanTest {
         val response2 = testedInterceptor.intercept(mockChain)
         val expectedResponse2 = fakeResponse
 
+        verify(localSpanBuilder).withOrigin(getExpectedOrigin())
         verify(localSpan).setTag("http.url", fakeUrl.lowercase(Locale.US))
         verify(localSpan).setTag("http.method", fakeMethod)
         verify(localSpan).setTag("http.status_code", statusCode)
         verify(localSpan, never()).finish()
+        verify(localSpan as MutableSpan).drop()
+        verify(mockSpanBuilder).withOrigin(getExpectedOrigin())
         verify(mockSpan).setTag("http.url", fakeUrl.lowercase(Locale.US))
         verify(mockSpan).setTag("http.method", fakeMethod)
         verify(mockSpan).setTag("http.status_code", statusCode)
         verify(mockSpan, never()).finish()
+        verify(mockSpan as MutableSpan).drop()
         assertThat(response1).isSameAs(expectedResponse1)
         assertThat(response2).isSameAs(expectedResponse2)
         verify(logger.mockInternalLogger)
@@ -1054,11 +1112,13 @@ internal open class TracingInterceptorNonDdTracerNotSendingSpanTest {
 
         val response = testedInterceptor.intercept(mockChain)
 
+        verify(mockSpanBuilder).withOrigin(getExpectedOrigin())
         verify(mockSpan).setTag("http.url", fakeUrl.lowercase(Locale.US))
         verify(mockSpan).setTag("http.method", fakeMethod)
         verify(mockSpan).setTag("http.status_code", statusCode)
         verify(mockSpan).setTag(tagKey, tagValue)
         verify(mockSpan, never()).finish()
+        verify(mockSpan as MutableSpan).drop()
         assertThat(response).isSameAs(fakeResponse)
     }
 
@@ -1080,11 +1140,13 @@ internal open class TracingInterceptorNonDdTracerNotSendingSpanTest {
 
         val response = testedInterceptor.intercept(mockChain)
 
+        verify(mockSpanBuilder).withOrigin(getExpectedOrigin())
         verify(mockSpan).setTag("http.url", fakeUrl.lowercase(Locale.US))
         verify(mockSpan).setTag("http.method", fakeMethod)
         verify(mockSpan).setTag("http.status_code", statusCode)
         verify(mockSpan).setTag(tagKey, tagValue)
         verify(mockSpan, never()).finish()
+        verify(mockSpan as MutableSpan).drop()
         assertThat(response).isSameAs(fakeResponse)
     }
 
@@ -1126,6 +1188,7 @@ internal open class TracingInterceptorNonDdTracerNotSendingSpanTest {
             testedInterceptor.intercept(mockChain)
         }
 
+        verify(mockSpanBuilder).withOrigin(getExpectedOrigin())
         verify(mockSpan).setTag("http.url", fakeUrl.lowercase(Locale.US))
         verify(mockSpan).setTag("http.method", fakeMethod)
         verify(mockSpan).setTag("error.type", throwable.javaClass.canonicalName)
@@ -1133,6 +1196,7 @@ internal open class TracingInterceptorNonDdTracerNotSendingSpanTest {
         verify(mockSpan).setTag("error.stack", throwable.loggableStackTrace())
         verify(mockSpan).setTag(tagKey, tagValue)
         verify(mockSpan, never()).finish()
+        verify(mockSpan as MutableSpan).drop()
     }
 
     @Test
@@ -1155,20 +1219,14 @@ internal open class TracingInterceptorNonDdTracerNotSendingSpanTest {
     @Test
     fun `M do not update the hostDetector W host list provided`(forge: Forge) {
         // GIVEN
-        val localHosts = forge.aList { forge.aStringMatching(HOSTNAME_PATTERN) }
+        val localHosts =
+            forge.aMap { forge.aStringMatching(HOSTNAME_PATTERN) to setOf(TracingHeaderType.DATADOG) }
 
         // WHEN
-        testedInterceptor =
-            instantiateTestedInterceptor(
-                localHosts.associateWith {
-                    setOf(TracingHeaderType.DATADOG)
-                }
-            ) {
-                mockLocalTracer
-            }
+        testedInterceptor = instantiateTestedInterceptor(localHosts) { mockLocalTracer }
 
         // THEN
-        verify(mockResolver, never()).addKnownHosts(localHosts)
+        verify(mockResolver, never()).addKnownHostsWithHeaderTypes(localHosts)
     }
 
     @Test
@@ -1246,11 +1304,11 @@ internal open class TracingInterceptorNonDdTracerNotSendingSpanTest {
 
     // region Internal
 
-    private fun stubChain(chain: Interceptor.Chain, statusCode: Int) {
+    internal fun stubChain(chain: Interceptor.Chain, statusCode: Int) {
         return stubChain(chain) { forgeResponse(statusCode) }
     }
 
-    private fun stubChain(
+    internal fun stubChain(
         chain: Interceptor.Chain,
         responseBuilder: () -> Response
     ) {
@@ -1303,9 +1361,7 @@ internal open class TracingInterceptorNonDdTracerNotSendingSpanTest {
     // endregion
 
     companion object {
-        const val HOSTNAME_PATTERN =
-            "(([a-zA-Z0-9]|[a-zA-Z0-9][a-zA-Z0-9\\-]{1,4}[a-zA-Z0-9]{2,3})\\.)+" +
-                "([A-Za-z]|[A-Za-z][A-Za-z0-9-]{1,2}[A-Za-z0-9])"
+        const val HOSTNAME_PATTERN = "([a-z][a-z0-9_~-]{3,9}\\.){1,4}[a-z][a-z0-9]{2,3}"
         const val IPV4_PATTERN =
             "(([0-9]|[1-9][0-9]|1[0-9]){2}\\.|(2[0-4][0-9]|25[0-5])\\.){3}" +
                 "([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])"
