+++ instrumented/integration/src/androidTest/kotlin/com/datadog/android/sdk/integration/rum/ResourceTrackingTest.kt
@@ -9,8 +9,8 @@ package com.datadog.android.sdk.integration.rum
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.filters.LargeTest
 import androidx.test.platform.app.InstrumentationRegistry
+import com.datadog.android.okhttp.rum.RumInterceptor
 import com.datadog.android.privacy.TrackingConsent
-import com.datadog.android.rum.RumInterceptor
 import com.datadog.android.rum.RumResourceAttributesProvider
 import com.datadog.android.sdk.integration.RuntimeConfig
 import com.datadog.android.sdk.rules.HandledRequest
