+++ instrumented/nightly-tests/build.gradle.kts
@@ -98,6 +98,7 @@ repositories {
 
 dependencies {
     implementation(project(":dd-sdk-android"))
+    implementation(project(":dd-sdk-android-okhttp"))
     implementation(project(":library:dd-sdk-android-ndk"))
     implementation(project(":library:dd-sdk-android-logs"))
